pyproject.toml
[tool.poetry]
name = "tasks"
version = "0.1.0"
description = "Managin project and tasks, tracking tasks"
authors = ["evetness <pap.kristof.199411@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
APIFlask = "^1.3.1"
marshmallow-dataclass = "^8.5.13"
pydantic = "^1.10.7"
sqlmodel = "^0.0.8"
alchemical = "^0.7.0"
Flask-Migrate = "^4.0.4"
Flask-Cors = "^3.0.10"
python-dotenv = "^1.0.0"
Flask-Caching = "^2.0.2"
redis = "^4.5.4"
loguru = "^0.7.0"
mariadb = "^1.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
